removed duplicate code
